Rê chuột vào đây đọc Description
Bản Update này đã hoàn thành MultiThread (tạm thời chạy chưa có lỗi)
+ Thêm rất rất nhiều dòng Comment cho code liên quan đến Server, Client, ReUpdateSystem, Updater(Client)
+ Bản game hiện tại có thể chạy trên wifi giữa 2 máy khác nhau (1 Client - 1 Server) nhiều hơn vẫn còn sai
- Chưa có comment Multi-Thread (quên)
- Còn nhiều phương thức đóng gói chưa viết
- Rốt cuộc là Client có bắt buộc phải có reUpdate như Server kooooooo? mô phỏng là cáo vẹo gì???????

+++ Server/Win32Project1/Client/Client.h
@@ -21,7 +21,7 @@ enum ePlayerFlags : uint8_t
 	player4
 };
 
-enum eObjectId : uint32_t
+enum eObjectId : uint8_t
 {
 	SizePack,
 	GameObject
@@ -51,7 +51,7 @@ enum eClientMode
 	Sending,
 };
 
-enum funcId : uint32_t
+enum funcId : uint8_t
 {
 	Pl_Move_Event,
 	Pl_Fire,
